# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backup management server for Postgres databases.",
        "title": "Backup Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/retention-policies": {
            "get": {
                "description": "Get a list of all configured retention policies (how many backups to keep)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "retention"
                ],
                "summary": "List all retention policies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.RetentionPolicy"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Define a new policy for automatic backup cleanup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "retention"
                ],
                "summary": "Create a new retention policy",
                "parameters": [
                    {
                        "description": "Retention policy configuration",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.CreateRetentionPolicyParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.RetentionPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/retention-policies/{id}": {
            "put": {
                "description": "Update a retention policy's configuration by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "retention"
                ],
                "summary": "Update an existing retention policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated configuration",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.UpdateRetentionPolicyParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.RetentionPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a retention policy configuration",
                "tags": [
                    "retention"
                ],
                "summary": "Delete a retention policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Policy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.CreateRetentionPolicyParams": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "keepDaily": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepHourly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepMonthly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepWeekly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepYearly": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "yearlyMonth": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.RetentionPolicy": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "id": {
                    "type": "string"
                },
                "keepDaily": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepHourly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepMonthly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepWeekly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepYearly": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "yearlyMonth": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.UpdateRetentionPolicyParams": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "keepDaily": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepHourly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepMonthly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepWeekly": {
                    "type": "integer",
                    "format": "int32"
                },
                "keepYearly": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "yearlyMonth": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}